Renamed and split trusted name implementation
