+++ src_features/provide_trusted_name/trusted_name.h
@@ -5,6 +5,8 @@
 
 #include <stdint.h>
 #include <stdbool.h>
+#include "common_utils.h"  // ADDRESS_LENGTH
+#include "tlv.h"
 
 #define TRUSTED_NAME_MAX_LENGTH 30
 
@@ -32,16 +34,42 @@ typedef enum {
     TN_SOURCE_COUNT,
 } e_name_source;
 
+typedef enum { TN_KEY_ID_DOMAIN_SVC = 0x07, TN_KEY_ID_CAL = 0x09 } e_tn_key_id;
+
+typedef struct {
+    bool valid;
+    uint8_t struct_version;
+    char *name;
+    uint8_t addr[ADDRESS_LENGTH];
+    uint64_t chain_id;
+    e_name_type name_type;
+    e_name_source name_source;
+    uint8_t nft_id[INT256_LENGTH];
+} s_trusted_name_info;
+
+typedef struct {
+    s_trusted_name_info *info;
+    e_tn_key_id key_id;
+    uint8_t input_sig_size;
+    uint8_t input_sig[73];
+    cx_sha256_t hash_ctx;
+    uint32_t rcv_flags;
+} s_trusted_name_ctx;
+
 const char *get_trusted_name(uint8_t type_count,
                              const e_name_type *types,
                              uint8_t source_count,
                              const e_name_source *sources,
                              const uint64_t *chain_id,
                              const uint8_t *addr);
-uint16_t handle_provide_trusted_name(uint8_t p1, const uint8_t *data, uint8_t length);
 
+extern s_trusted_name_info g_trusted_name_info;
 extern char g_trusted_name[TRUSTED_NAME_MAX_LENGTH + 1];
 
-#endif  // TRUSTED_NAME_H_
+bool handle_trusted_name_struct(const s_tlv_data *data, s_trusted_name_ctx *context);
+bool verify_trusted_name_struct(const s_trusted_name_ctx *ctx);
+bool verify_trusted_name_signature(const s_trusted_name_ctx *ctx);
 
-#endif  // HAVE_TRUSTED_NAME
+#endif  // !TRUSTED_NAME_H_
+
+#endif  // !HAVE_TRUSTED_NAME
